% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift4gains2.R
\name{lift4gains2}
\alias{lift4gains2}
\title{Functions for plotting cumulative gains, lift and response charts}
\usage{
lift4gains2(H2OAutoML_object, response_ref = NULL, save_pngs = F)
}
\arguments{
\item{H2OAutoML_object}{provided by user}

\item{response_ref}{provided by user}

\item{save_pngs}{provided by user}
}
\description{
Functions for plotting cumulative gains, lift and response charts
}
