% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curves.R
\name{roc_curves}
\alias{roc_curves}
\title{Functions for plotting roc curves}
\usage{
roc_curves(H2OAutoML_object, plot = T, best = F, save_png = F)
}
\arguments{
\item{H2OAutoML_object}{provided by user}
}
\description{
Functions for plotting roc curves
}
