% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp_ggplot.R
\name{varImp_ggplot}
\alias{varImp_ggplot}
\title{Functions for plotting variable importance}
\usage{
varImp_ggplot(H2OAutoML_object, save_pngs = F, return_data = F)
}
\arguments{
\item{H2OAutoML_object}{provided by user}

\item{save_pngs}{provided by user}

\item{return_data}{provided by user}
}
\description{
Functions for plotting variable importance
}
