% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{varImp_plot}
\alias{varImp_plot}
\title{Functions for plotting variable importance}
\usage{
varImp_plot(H2OAutoML_object, save_pngs = F, return_data = F)
}
\arguments{
\item{H2OAutoML_object}{provided by user}
}
\description{
Functions for plotting variable importance
}
