% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_bars.R
\name{auc_bars}
\alias{auc_bars}
\title{Function for plotting AUC bars}
\usage{
auc_bars(H2OAutoML_object, save_png = F, test_data)
}
\arguments{
\item{H2OAutoML_object}{An object containing multiple models trained in H2O.}

\item{save_png}{(Optional) Whether to save a .png with ggsave(). Default is FALSE.}
}
\description{
Function for plotting AUC bars
}
