% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsLiftResponse.R
\name{lift4gains}
\alias{lift4gains}
\title{Functions for plotting cumulative gains, lift and response charts}
\usage{
lift4gains(H2OAutoML_object, response_ref = NULL, save_pngs = F)
}
\arguments{
\item{H2OAutoML_object}{provided by user}
}
\description{
Functions for plotting cumulative gains, lift and response charts
}
