% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curves.R
\name{roc_curves}
\alias{roc_curves}
\title{Functions for plotting roc curves}
\usage{
roc_curves(H2OAutoML_object, plot = T, best = F, save_png = F,
  test_data)
}
\arguments{
\item{H2OAutoML_object}{An object containing multiple models trained in H2O.}

\item{plot}{(Optional) Whether to plot the output. Default is TRUE.}

\item{best}{(Optional) If set to TRUE, the function returns only the best model.}

\item{save_png}{(Optional) Whether to save a .png with ggsave(). Default is FALSE.}

\item{test_data}{The test data set.}
}
\description{
Functions for plotting roc curves
}
